+++ order/src/main/java/com/sparta/hotdeal/order/presentation/dtos/basket/req/ReqPatchBasketDto.java
@@ -1,16 +1,14 @@
-package com.sparta.hotdeal.order.presentation.dtos.order;
+package com.sparta.hotdeal.order.presentation.dtos.basket.req;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import lombok.ToString;
 
 @Getter
 @Builder
-@ToString
 @NoArgsConstructor
 @AllArgsConstructor
-public class ReqPatchOrderDto {
-    private String orderStatus;
+public class ReqPatchBasketDto {
+    private Integer quantity;
 }
