+++ order/src/main/java/com/sparta/hotdeal/order/presentation/dtos/basket/res/ResGetBasketByIdDto.java
@@ -1,4 +1,4 @@
-package com.sparta.hotdeal.order.presentation.dtos.basket;
+package com.sparta.hotdeal.order.presentation.dtos.basket.res;
 
 import java.util.UUID;
 import lombok.AllArgsConstructor;
