+++ order/src/main/java/com/sparta/hotdeal/order/presentation/dtos/order/req/ReqPatchOrderDto.java
@@ -1,14 +1,16 @@
-package com.sparta.hotdeal.order.presentation.dtos.basket;
+package com.sparta.hotdeal.order.presentation.dtos.order.req;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
+import lombok.ToString;
 
 @Getter
 @Builder
+@ToString
 @NoArgsConstructor
 @AllArgsConstructor
-public class ReqPatchBasketDto {
-    private Integer quantity;
+public class ReqPatchOrderDto {
+    private String orderStatus;
 }
