+++ order/src/main/java/com/sparta/hotdeal/order/presentation/dtos/order/res/ResGetOrderByIdDto.java
@@ -1,4 +1,4 @@
-package com.sparta.hotdeal.order.presentation.dtos.order;
+package com.sparta.hotdeal.order.presentation.dtos.order.res;
 
 import java.time.LocalDate;
 import java.util.Arrays;
